+++ app/build.gradle
@@ -165,4 +165,7 @@ dependencies {
     // https://mvnrepository.com/artifact/org.apache.poi/poi-ooxml
     implementation("org.apache.poi:poi-ooxml:5.2.5")
 
+    implementation ("org.zeroturnaround:zt-zip:1.17")
+
+
 }
